회원 탈퇴

+++ data/src/main/java/com/packy/data/model/auth/SignUpDto.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class SignUpDto(
     @SerialName("accessToken") val accessToken: String,
-    @SerialName("accessTokenExpiresIn") val accessTokenExpiresIn: Int,
+    @SerialName("accessTokenExpiresIn") val accessTokenExpiresIn: Long,
     @SerialName("grantType") val grantType: String,
     @SerialName("refreshToken") val refreshToken: String
 )
